Finished refactor and comments.
